requirements.txt
# API и асинхронная работа
fastapi==0.103.1
uvicorn==0.23.2
gunicorn==21.2.0
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6

# Безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Базы данных
sqlalchemy==2.0.21
alembic==1.12.0
asyncpg==0.28.0
psycopg2-binary==2.9.9
redis==4.6.0
aioredis==2.0.1
greenlet==2.0.2

# Парсинг файлов
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1

# Работа с ML/прогнозами
# prophet устанавливается через conda: conda install -c conda-forge prophet
pmdarima==2.0.4
scikit-learn==1.3.1
statsmodels==0.14.0

# Работа со штрих-кодами
opencv-python==4.8.0.74
pyzbar==0.1.9
Pillow==10.0.0

# Генерация ID
nanoid==2.0.0

# Задачи и очереди
celery==5.3.4
python-dotenv==1.0.0

# Email
fastapi-mail==1.4.1

# Тесты
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0

# Логирование
loguru==0.7.0

# Мониторинг
prometheus-client==0.17.1